pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "End-to-End Q&A Chatbot using LangChain and OpenAI"
readme = "README.md"
requires-python = ">=3.10,<3.13"

dependencies = [
    "langchain>=0.3.26",
    "langchain-openai>=0.1.7",
    "llama-index>=0.11.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.34.0",
    "langsmith>=0.1.20",
    "langchain-community>=0.0.38",
    "langchain-core>=0.1.60",
    "langserve>=0.0.45",
    "langchain-groq>=0.3.8",
    "google-search-results>=2.4.3",
    "gtts>=2.5.1",
    "SpeechRecognition>=3.10.1",
    "faiss-cpu>=1.8.0",
    "docx2txt>=0.8",
]
